add dinamic data for modal and main page
