+++ src/App.js
@@ -1,34 +1,21 @@
-import React, { lazy, useState } from 'react';
+import React, { lazy } from 'react';
 import { RestrictedRoute } from 'components/RestrictedRoute';
 import { Route, Routes } from 'react-router-dom';
 import { GlobalStyle } from 'components/GlobalStyle';
-import { ModalDailyCalorie } from 'components/ModalDailyCalorie/ModalDailyCalorie';
-import { DailyCaloriesForm } from 'components/DailyCaloriesForm';
 import { DiaryDateCalendar } from 'components/DiaryDateCalendar/DiaryDateCalendar';
 
-// import { ModalDailyCalorie } from 'components/ModalDailyCalorie/ModalDailyCalorie';
-// import { DairyProductList } from 'components/DairyProductList/DairyProductList';
-// import { MainPage } from 'components/DailyCaloriesForm';
+import { MainPage } from 'pages/MainPage';
 
 const RegistrationPage = lazy(() => import('pages/RegistrationPage'));
 const LoginPage = lazy(() => import('pages/LoginPage'));
 
 export const App = () => {
-  const [isModalOpen, setIsModalopen] = useState(false);
-  const closeModal = () => {
-    setIsModalopen(false);
-    window.document.body.style.overflow = 'unset';
-  };
-  const openModal = () => {
-    setIsModalopen(true);
-    window.document.body.style.overflow = 'hidden';
-  };
   return (
     <>
       <GlobalStyle />
-      {isModalOpen && <ModalDailyCalorie closeModal={closeModal} />}
-      {/* <DailyCaloriesForm /> */}
       <Routes>
+        <Route path="/" element={<MainPage />} />
+
         <Route
           path="/signup"
           element={
